feat(entities): 추가된 테스트 실행 및 테스트 케이스 실행 유형 정의 리팩터링
- TestRunDetail, TestCaseRunDetail, UpdateTestCaseRunInput 등 새로운 유형 정의 추가
- features에서 사용되는 중복된 타입 entities로 이동 및 참조 업데이트
- 코드 중복 제거 및 데이터 타입 관리 통합
