+++ src/entities/test-run/index.ts
@@ -1,2 +1,6 @@
 export { TestRunStatusEnum, TestRunSchema, CreateTestRunSchema } from './model';
-export type { TestRunDTO, CreateTestRunDTO, CreateTestRunInput, TestRunStatus, TestRunSourceType } from './model';
+export type {
+  TestRunDTO, CreateTestRunDTO, CreateTestRunInput, TestRunStatus, TestRunSourceType,
+  TestCaseRunDetail, SourceInfo, TestRunDetail, FetchedTestRun,
+  UpdateTestCaseRunInput, UpdateTestCaseRunResult,
+} from './model';
