+++ src/entities/test-run/model/index.ts
@@ -1,2 +1,6 @@
 export { TestRunStatusEnum, TestRunSchema, CreateTestRunSchema } from './schema';
-export type { TestRunDTO, CreateTestRunDTO, CreateTestRunInput, TestRunStatus, TestRunSourceType } from './types';
+export type {
+  TestRunDTO, CreateTestRunDTO, CreateTestRunInput, TestRunStatus, TestRunSourceType,
+  TestCaseRunDetail, SourceInfo, TestRunDetail, FetchedTestRun,
+  UpdateTestCaseRunInput, UpdateTestCaseRunResult,
+} from './types';
